Run Cucumber features from `gem test cape`
Rework Cucumber features to avoid depending on Bundler. This is so that
the installed gem can also run features. Making Bundler a test
dependency is not an option because our .gemspec uses Git to find our
source files, and the installed gem is not a Git repository. (Nor would
we want to make Git a test dependency.)