Warn instead of bubbling Capistrano's ArgumentError on illegal task name
The readme implies you can mirror all Rake tasks in a standard Rails
project by simply calling Cape.mirror_rake_tasks with no arguments.
Capistrano raises ArgumentError when you try to define a task whose name
collides with a Ruby method name. For example a 'load' Rake task
collides with the core method Kernel#load when you try to mirror it as a
Capistrano recipe. There are Rake tasks defined by Rails that have the
same problem.

Instead of bubbling the error from Capistrano, we are writing a warning
to stderr and ignoring the offending task.